requirements.txt
influxdb_client
pyyaml
bs4
ujson